pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.6",
    "greenlet>=3.3.2",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "pandas>=3.0.1",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
